+++ core/thread.py
@@ -10,14 +10,13 @@ from datetime import datetime, timedelta
 import discord
 from discord.ext.commands import MissingRequiredArgument, CommandError
 
-from core.models import Bot, ThreadManagerABC, ThreadABC
 from core.utils import is_image_url, days, match_user_id
 from core.utils import truncate, ignore, error
 
 logger = logging.getLogger('Modmail')
 
 
-class Thread(ThreadABC):
+class Thread:
     """Represents a discord Modmail thread"""
 
     def __init__(self, manager: 'ThreadManager',
@@ -42,36 +41,36 @@ class Thread(ThreadABC):
         return (f'Thread(recipient="{self.recipient or self.id}", '
                 f'channel={self.channel.id})')
 
-    async def wait_until_ready(self):
+    async def wait_until_ready(self) -> None:
         """Blocks execution until the thread is fully set up."""
         await self._ready_event.wait()
 
     @property
-    def id(self):
+    def id(self) -> int:
         return self._id
 
     @property
-    def close_task(self):
+    def close_task(self) -> asyncio.TimerHandle:
         return self._close_task
 
     @close_task.setter
-    def close_task(self, val):
+    def close_task(self, val: asyncio.TimerHandle):
         self._close_task = val
 
     @property
-    def channel(self):
+    def channel(self) -> typing.Union[discord.TextChannel, discord.DMChannel]:
         return self._channel
 
     @property
-    def recipient(self):
+    def recipient(self) -> typing.Optional[typing.Union[discord.User, discord.Member]]:
         return self._recipient
 
     @property
-    def ready(self):
+    def ready(self) -> bool:
         return self._ready_event.is_set()
 
     @ready.setter
-    def ready(self, flag):
+    def ready(self, flag: bool):
         if flag:
             self._ready_event.set()
         else:
@@ -94,7 +93,7 @@ class Thread(ThreadABC):
             overwrites = None
 
         channel = await self.bot.modmail_guild.create_text_channel(
-            name=self.manager._format_channel_name(recipient),
+            name=self.manager.format_channel_name(recipient),
             category=category,
             overwrites=overwrites,
             reason='Creating a thread channel'
@@ -114,9 +113,9 @@ class Thread(ThreadABC):
             log_url = log_count = None
             # ensure core functionality still works
 
-        info_embed = self.manager._format_info_embed(recipient, log_url,
-                                                     log_count,
-                                                     discord.Color.green())
+        info_embed = self.manager.format_info_embed(recipient, log_url,
+                                                    log_count,
+                                                    discord.Color.green())
 
         topic = f'User ID: {recipient.id}'
         if creator:
@@ -169,8 +168,11 @@ class Thread(ThreadABC):
             self._close(closer, silent, delete_channel, message, True)
         )
 
-    async def close(self, *, closer, after=0, silent=False,
-                    delete_channel=True, message=None):
+    async def close(self, *, closer: typing.Union[discord.Member, discord.User],
+                    after: int = 0,
+                    silent: bool = False,
+                    delete_channel: bool = True,
+                    message: str = None) -> None:
         """Close a thread now or after a set time in seconds"""
 
         # restarts the after timer
@@ -304,7 +306,7 @@ class Thread(ThreadABC):
 
         await asyncio.gather(*tasks)
 
-    async def cancel_closure(self):
+    async def cancel_closure(self) -> None:
         if self.close_task is not None:
             self.close_task.cancel()
             self.close_task = None
@@ -323,7 +325,8 @@ class Thread(ThreadABC):
                 if str(message_id) == str(embed.author.url).split('/')[-1]:
                     return msg
 
-    async def edit_message(self, message_id, message):
+    async def edit_message(self, message_id: typing.Union[int, str],
+                           message: str) -> None:
         msg_recipient, msg_channel = await asyncio.gather(
             self._find_thread_message(self.recipient, message_id),
             self._find_thread_message(self.channel, message_id)
@@ -348,7 +351,7 @@ class Thread(ThreadABC):
             msg_channel.delete()
         )
 
-    async def note(self, message):
+    async def note(self, message: discord.Message) -> None:
         if not message.content and not message.attachments:
             raise MissingRequiredArgument(param(name='msg'))
 
@@ -359,7 +362,8 @@ class Thread(ThreadABC):
             self.send(message, self.channel, note=True)
         )
 
-    async def reply(self, message, anonymous=False):
+    async def reply(self, message: discord.Message,
+                    anonymous: bool = False) -> None:
         if not message.content and not message.attachments:
             raise MissingRequiredArgument(param(name='msg'))
         if all(not g.get_member(self.id) for g in self.bot.guilds):
@@ -417,8 +421,12 @@ class Thread(ThreadABC):
 
         await asyncio.gather(*tasks)
 
-    async def send(self, message, destination=None,
-                   from_mod=False, note=False, anonymous=False):
+    async def send(self, message: discord.Message,
+                   destination: typing.Union[discord.TextChannel, discord.DMChannel,
+                                             discord.User, discord.Member] = None,
+                   from_mod: bool = False,
+                   note: bool = False,
+                   anonymous: bool = False) -> None:
         if self.close_task is not None:
             # cancel closing if a thread message is sent.
             self.bot.loop.create_task(
@@ -569,7 +577,7 @@ class Thread(ThreadABC):
         if delete_message:
             self.bot.loop.create_task(ignore(message.delete()))
 
-    def get_notifications(self):
+    def get_notifications(self) -> str:
         config = self.bot.config
         key = str(self.id)
 
@@ -584,14 +592,14 @@ class Thread(ThreadABC):
         return ' '.join(mentions)
 
 
-class ThreadManager(ThreadManagerABC):
+class ThreadManager:
     """Class that handles storing, finding and creating Modmail threads."""
 
-    def __init__(self, bot: Bot):
+    def __init__(self, bot):
         self.bot = bot
         self.cache = {}
 
-    async def populate_cache(self):
+    async def populate_cache(self) -> None:
         for channel in self.bot.modmail_guild.text_channels:
             if channel.category != self.bot.main_category and not \
                     self.bot.using_multiple_server_setup:
@@ -604,10 +612,13 @@ class ThreadManager(ThreadManagerABC):
     def __iter__(self):
         return iter(self.cache.values())
 
-    def __getitem__(self, item):
+    def __getitem__(self, item: str) -> Thread:
         return self.cache[item]
 
-    async def find(self, *, recipient=None, channel=None, recipient_id=None):
+    async def find(self, *,
+                   recipient: typing.Union[discord.Member, discord.User] = None,
+                   channel: discord.TextChannel = None,
+                   recipient_id: int = None) -> Thread:
         """Finds a thread from cache or from discord channel topics."""
         if recipient is None and channel is not None:
             return await self._find_from_channel(channel)
@@ -672,7 +683,9 @@ class ThreadManager(ThreadManagerABC):
 
             return thread
 
-    def create(self, recipient, *, creator=None, category=None):
+    def create(self, recipient: typing.Union[discord.Member, discord.User], *,
+               creator: typing.Union[discord.Member, discord.User] = None,
+               category: discord.CategoryChannel = None) -> Thread:
         """Creates a Modmail thread"""
         # create thread immediately so messages can be processed
         thread = Thread(self, recipient)
@@ -685,10 +698,12 @@ class ThreadManager(ThreadManagerABC):
         ))
         return thread
 
-    async def find_or_create(self, recipient):
+    async def find_or_create(self,
+                             recipient: typing.Union[discord.Member,
+                                                     discord.User]) -> Thread:
         return await self.find(recipient=recipient) or self.create(recipient)
 
-    def _format_channel_name(self, author):
+    def format_channel_name(self, author):
         """Sanitises a username for use with text channel names"""
         name = author.name.lower()
         allowed = string.ascii_letters + string.digits + '-'
@@ -701,7 +716,7 @@ class ThreadManager(ThreadManagerABC):
 
         return new_name
 
-    def _format_info_embed(self, user, log_url, log_count, color):
+    def format_info_embed(self, user, log_url, log_count, color):
         """Get information about a member of a server
         supports users from the guild or not."""
         member = self.bot.guild.get_member(user.id)
